+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog
 Version 0.2.3
 =============
 * ✨ Added ``group`` command
+* 🚑 Fixed a bug where pass-through flags were being mangled
 
 `Version 0.2.2: Autobind <https://github.com/shnizzedy/cpac-python-package/releases/tag/v0.2.2>`_
 =================================================================================================